feat: separate mcp function
